refactor(rest): extract storage/asset routes into routes/storage.py
Moves the 12 user-facing storage routes (files, overlays, blobs
GET/PUT/DELETE, keys/move-to-folder, keys/rename, derived, upload-url,
upload-complete, upload-progress, download-url) out of create_app's
closure into routes/storage.py.

Lifts the pure helpers these routes need onto routes/deps.py:
is_accepted_source, format_label/SOURCE_FORMAT_NAMES,
content_encoding_for/GZIP_UPLOAD_EXTS, parse_move_body,
parse_rename_body, UPLOAD_URL_TTL_SECONDS. create_app keeps the old
underscore-prefixed names bound to them for the routes still in the
closure (admin storage list/compression, convert, admin key
move/rename). rename_with_status is exported from routes/storage.py
itself (rather than deps.py) since it is storage-cascade logic
specific to this group, and is called directly by the admin rename
route until that route is extracted too.
